New communication protocols

+++ vesc_uart/buffer.c
@@ -90,7 +90,7 @@ void buffer_append_float32_auto(uint8_t* buffer, float number, int32_t *index) {
 
 	uint32_t res = ((e & 0xFF) << 23) | (sig_i & 0x7FFFFF);
 	if (sig < 0) {
-		res |= 1 << 31;
+		res |= (uint32_t)(1 << 31);
 	}
 
 	buffer_append_uint32(buffer, res, index);
@@ -171,7 +171,7 @@ float buffer_get_float32_auto(const uint8_t *buffer, int32_t *index) {
 
 	int e = (res >> 23) & 0xFF;
 	uint32_t sig_i = res & 0x7FFFFF;
-	bool neg = res & (1 << 31);
+	bool neg = res & (uint32_t)(1 << 31);
 
 	float sig = 0.0;
 	if (e != 0 || sig_i != 0) {
